List available job type versions
