+++ image_builder/image_builder/docker/builder.py
@@ -1,15 +1,14 @@
-from collections import defaultdict
-from dataclasses import dataclass
 import os
 from pathlib import Path
 import time
-from typing import List, Optional, Dict, Tuple
+from typing import Optional, Dict, Tuple
 
 import backoff
 
 from image_builder.base import ImageBuilder
 from image_builder.config import Config
 from image_builder.docker.template import template_dockerfile
+from image_builder.job_type import JobTypeVersion, load_job_type
 from image_builder.metrics import (
     metric_images_built,
     metric_images_building_errors,
@@ -19,22 +18,12 @@ from image_builder.metrics import (
 from racetrack_client.log.context_error import wrap_context, ContextError
 from racetrack_client.log.logs import get_logger
 from racetrack_client.manifest import Manifest
-from racetrack_client.utils.semver import SemanticVersion
 from racetrack_client.utils.shell import shell, CommandError, shell_output
 from racetrack_client.utils.url import join_paths
 from racetrack_commons.deploy.image import get_fatman_image, get_fatman_user_module_image
-from racetrack_commons.plugin.core import PluginCore
 from racetrack_commons.plugin.engine import PluginEngine
 
 logger = get_logger(__name__)
-
-
-@dataclass
-class JobTypeVersion:
-    lang_name: str
-    version: str  # semantic version of the job type
-    base_image_path: Path  # base Dockerfile path
-    template_path: Path  # fatman template Dockerfile path
 
 
 class DockerBuilder(ImageBuilder):
@@ -50,7 +39,7 @@ class DockerBuilder(ImageBuilder):
         plugin_engine: PluginEngine,
     ) -> Tuple[str, str, Optional[str]]:
         """Build image from manifest file in a workspace directory and return built image name"""
-        job_type: JobTypeVersion = _load_job_type(config, plugin_engine, manifest.lang)
+        job_type: JobTypeVersion = load_job_type(plugin_engine, manifest.lang)
 
         _wait_for_docker_engine_ready()
 
@@ -163,55 +152,3 @@ def _image_exists_in_registry(image_name: str):
 @backoff.on_exception(backoff.fibo, CommandError, max_value=3, max_time=30, jitter=None)
 def _wait_for_docker_engine_ready():
     shell('docker ps')
-
-
-@backoff.on_exception(backoff.fibo, AssertionError, max_value=1, max_time=5, jitter=None, logger=None)
-def _load_job_type(
-    config: Config,
-    plugin_engine: PluginEngine,
-    lang: str,
-) -> JobTypeVersion:
-    """
-    Load job type.
-    In case it's not found, retry attempts due to possible delayed plugins loading
-    """
-    with wrap_context('gathering available job types'):
-        job_types = _gather_job_types(config, plugin_engine)
-    assert job_types, f'language {lang} is not supported. Extend Racetrack with job type plugins'
-    assert lang in job_types, f'language {lang} is not supported, supported are: {sorted(job_types.keys())}'
-    return job_types[lang]
-
-
-def _gather_job_types(
-    config: Config,
-    plugin_engine: PluginEngine,
-) -> Dict[str, JobTypeVersion]:
-    """
-    Load job types from plugins.
-    Return job name (with version) -> (base image name, dockerfile template path)
-    """
-    job_types: Dict[str, JobTypeVersion] = {}
-    job_family_versions: Dict[str, List[JobTypeVersion]] = defaultdict(list)
-
-    plugin_results: List[Dict[str, Tuple[Path, Path]]] = plugin_engine.invoke_plugin_hook(PluginCore.fatman_job_types)
-    for plugin_job_types in plugin_results:
-        if plugin_job_types:
-            for job_full_name, job_data in plugin_job_types.items():
-                base_image_path, template_path = job_data
-                assert base_image_path.is_file(), f'cannot find base image Dockerfile for {job_full_name} language wrapper: {base_image_path}'
-                assert template_path.is_file(), f'cannot find Dockerfile template for {job_full_name} language wrapper: {template_path}'
-                name_parts = job_full_name.split(':')
-                assert len(name_parts) == 2, f'job type {job_full_name} should have the version defined (name:version)'
-                lang_name = name_parts[0]
-                lang_version = name_parts[1]
-                job_type_version = JobTypeVersion(lang_name, lang_version, base_image_path, template_path)
-                job_types[job_full_name] = job_type_version
-                job_family_versions[lang_name].append(job_type_version)
-
-    for lang_name in job_family_versions.keys():
-        versions = job_family_versions[lang_name]
-        versions.sort(key=lambda v: SemanticVersion(v.version))
-        latest_version = versions[-1]
-        job_types[f'{lang_name}:latest'] = latest_version
-
-    return job_types
